+++ modules/kubernetes/aks-core/modules.tf
@@ -1,6 +1,7 @@
 locals {
   exclude_namespaces = [
     "aad-pod-identity",
+    "azad-kube-proxy",
     "azdo-proxy",
     "calico-system",
     "cert-manager",
@@ -58,11 +59,12 @@ module "fluxcd_v2_azure_devops" {
   namespaces = [for ns in var.namespaces : {
     name = ns.name
     flux = {
-      enabled     = ns.flux.enabled
-      create_crds = ns.flux.create_crds
-      org         = ns.flux.azure_devops.org
-      proj        = ns.flux.azure_devops.proj
-      repo        = ns.flux.azure_devops.repo
+      enabled             = ns.flux.enabled
+      create_crds         = ns.flux.create_crds
+      include_tenant_name = ns.flux.include_tenant_name
+      org                 = ns.flux.azure_devops.org
+      proj                = ns.flux.azure_devops.proj
+      repo                = ns.flux.azure_devops.repo
     }
   }]
 }
@@ -495,13 +497,12 @@ module "promtail" {
   }
 
   source              = "../../kubernetes/promtail"
-  cloud_provider      = "azure"
   cluster_name        = "${var.name}${local.aks_name_suffix}"
   environment         = var.environment
   region              = var.location_short
   excluded_namespaces = var.promtail_config.excluded_namespaces
-
-  loki_address = var.promtail_config.loki_address
+  cluster_id          = local.cluster_id
+  loki_address        = var.promtail_config.loki_address
   azure_config = {
     azure_key_vault_name = var.promtail_config.azure_key_vault_name
     identity             = var.promtail_config.identity